refactor: aws capability config 통합 - 기존에 aws 자원별로 분산되어있던 capability config 통합
